Refactor multi-sandbox support into multi.js
